+++ __tests__/node/repositories/messageRespository.test.ts
@@ -1,27 +1,27 @@
-import {MessageRepository} from "@/repositories/message/respository";
-import {Redis} from "@upstash/redis";
+import { MessageRepository } from "@/repositories/message/repository";
+import { Redis } from "@upstash/redis";
 
-describe('repository.sendMessage tests',()=>{
+describe('repository.sendMessage tests', () => {
     let mockDb: Redis
     let repository: MessageRepository
-    beforeEach(()=>{
+    beforeEach(() => {
         mockDb = {
             zadd: jest.fn()
         } as unknown as Redis
-        repository = new MessageRepository(mockDb )
+        repository = new MessageRepository(mockDb)
     })
-    it('zadd should be called on database',async ()=>{
-        await repository.sendMessage('chatId', {id: 'id', senderId: 'senderId', text: 'text', timestamp: 123})
+    it('zadd should be called on database', async () => {
+        await repository.sendMessage('chatId', { id: 'id', senderId: 'senderId', text: 'text', timestamp: 123 })
         expect(mockDb.zadd).toHaveBeenCalled()
     })
-    it('check arguments passed to zadd',async ()=>{
-        await repository.sendMessage('chatId', {id: 'id', senderId: 'senderId', text: 'text', timestamp: 123})
-        expect(mockDb.zadd).toHaveBeenCalledWith( "chat:chatId:messages", {"member": "{\"id\":\"id\",\"senderId\":\"senderId\",\"text\":\"text\",\"timestamp\":123}", "score": 123})
+    it('check arguments passed to zadd', async () => {
+        await repository.sendMessage('chatId', { id: 'id', senderId: 'senderId', text: 'text', timestamp: 123 })
+        expect(mockDb.zadd).toHaveBeenCalledWith("chat:chatId:messages", { "member": "{\"id\":\"id\",\"senderId\":\"senderId\",\"text\":\"text\",\"timestamp\":123}", "score": 123 })
     })
 })
 
-describe('repository.removeAllMessages tests',()=>{
-    it('check arguments passed to database.del',()=>{
+describe('repository.removeAllMessages tests', () => {
+    it('check arguments passed to database.del', () => {
         const mockDb: Redis = {
             del: jest.fn()
         } as unknown as Redis
