+++ src/controllers/message/abstractController.ts
@@ -1,7 +1,7 @@
-import {AbstractController} from "@/controllers/abstractController";
-import {MessageRepository} from "@/repositories/message/respository";
+import { AbstractController } from "@/controllers/abstractController";
+import { MessageRepository } from "@/repositories/message/repository";
 
-export abstract class AbstractMessageController extends AbstractController{
+export abstract class AbstractMessageController extends AbstractController {
     protected readonly messageRepository: MessageRepository
     constructor() {
         super();
